+++ CHANGELOG.rst
@@ -17,9 +17,13 @@ Unreleased
 ----------
 * nothing unreleased
 
+[4.20.10]
+---------
+* fix: fixing the status and datetime of removed group memberships
+
 [4.20.9]
 --------
-* fix: fixing the status and datetime of removed group memberships
+* feat: Update permissions to grant access to provisioning admins
 
 [4.20.8]
 --------